% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agedata.R
\docType{data}
\name{agedata}
\alias{agedata}
\title{agedata}
\format{
R data frame
}
\usage{
agedata
}
\description{
A data frame of peoples' age
\itemize{
\item \code{age} - age in years
}
}
\keyword{datasets}
