% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BR_LatSq.R
\docType{data}
\name{BR_LatSq}
\alias{BR_LatSq}
\title{BR_LatSq}
\format{
R data frame.
}
\usage{
BR_LatSq
}
\description{
\code{BR_LatSq} contains data from Table 3 of Bliss, C.I. and Rose, C.L. (1940).
\itemize{
\item \code{Dogs} - Dog number
\item \code{Weeks} - Week number
\item \code{Treat} - Treatment
\item \code{value} - Mg-per cent serum calcium
}
}
\keyword{datasets}
