% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemplant.R
\docType{data}
\name{chemplant}
\alias{chemplant}
\title{chemplant}
\format{
R data frame.
}
\usage{
chemplant
}
\description{
\code{chemplant} contains data from section 5.2 of Box et al. (2005)
\itemize{
\item \code{run} - run number
\item \code{T} - Level of temperature
\item \code{C} - Level of concentration
\item \code{K} - Level of catalyst
\item \code{y} - Yield
}
}
\keyword{datasets}
