% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\docType{data}
\name{cookies}
\alias{cookies}
\title{cookies}
\format{
R data frame.
}
\usage{
cookies
}
\description{
\code{chemplant} contains data from section 5.2 of Box et al. (2005)
\itemize{
\item \code{Run} - run number
\item \code{Butter} - Level of butter
\item \code{Sugar} - Level of sugar
\item \code{Powder} - Level of baking powder
\item \code{Taste} - Taste on a scale of 1 to 10
}
}
\keyword{datasets}
