% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_trial.R
\docType{data}
\name{covid19_trial}
\alias{covid19_trial}
\title{covid19_trial}
\format{
R data frame.
}
\usage{
covid19_trial
}
\description{
\code{covid19_trial} is simulated data based on \href{https://clinicaltrials.gov/ct2/show/NCT04358068}{Evaluating the efficacy of hydroxychloroquine and azithromycin to prevent hospitalization or death in persons with covid-19}
\itemize{
\item \code{patient} - Patient ID
\item \code{trt} - Treatment indicator
\item \code{die_hosp} - Died in hospital (1 = yes, 0 = no)
\item \code{age} - Age of patient (years)
}
}
\keyword{datasets}
