% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSectdat.R
\docType{data}
\name{CSectdat}
\alias{CSectdat}
\title{CSectdat}
\format{
R data frame.
}
\usage{
CSectdat
}
\description{
\code{CSectdat} is data from \href{https://www.obgyn.utoronto.ca/hospital-statistics}{Obstetrics and Gynaecology, University of Toronto}
\itemize{
\item \code{Hospital} - \code{MG} or \code{NYG}
\item \code{CSecthosp} - Indicator of C-section (1 = yes, 0 = no)
}
}
\keyword{datasets}
