% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertdat.R
\docType{data}
\name{fertdat}
\alias{fertdat}
\title{fertdat}
\format{
R data frame.
}
\usage{
fertdat
}
\description{
\code{fertdat} is data on the effect of two treatments on wheat yield
\itemize{
\item \code{trt} - Level of fertilizer (A or B)
\item \code{fert} - wheat yield
}
}
\keyword{datasets}
