% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvdat.R
\docType{data}
\name{hsvdat}
\alias{hsvdat}
\title{hsvdat}
\format{
R data frame.
}
\usage{
hsvdat
}
\description{
\code{hsvdat} is data from Jaynes et al. (2013). Application of fractional factorial designs to study drug combinations

The columns are:

\code{Run} - Run number

\code{A}, \code{B}, \code{C}, \code{D}, \code{E}, \code{F} - Indicator of drug level (-1 = low, +1 = high)

\code{readout} - percentage of cells positive for HSV-1 after therapy
}
\keyword{datasets}
