% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafspring.R
\docType{data}
\name{leafspring}
\alias{leafspring}
\title{leafspring}
\format{
R data frame.
}
\usage{
leafspring
}
\description{
\code{leafspring} is data from Section 5.1 of Wu and Hammada (2011)

The columns are:

\code{B}, \code{C}, \code{D}, \code{E}, \code{Q} - Indicator of factor level (-1 = low, +1 = high)

\code{y1}, \code{y2}, \code{y3} - three replications of free height measurement
}
\keyword{datasets}
