% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifesat_childmort.R
\docType{data}
\name{lifesat_childmort}
\alias{lifesat_childmort}
\title{lifesat_childmort}
\format{
R data frame.
}
\usage{
lifesat_childmort
}
\description{
\code{lifesat_childmort} data from \href{https://worldhappiness.report/ed/2019/}{World Happiness Report 2019}

The columns are:

\code{Country} - Country

\verb{Country Code} - Country code

\code{Year} - Year

\code{Tot_pop} - Total poulation

\code{Continent} - Continent

\code{LifeSatisfaction} - Life satisfaction from Cantril Ladder (10 is the best and 0 is the worst)

\code{Under5mort} - Under 5 mortality per 1000
}
\keyword{datasets}
