% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhefs9282.R
\docType{data}
\name{nhefs9282}
\alias{nhefs9282}
\title{nhefs9282}
\format{
R data frame.
}
\usage{
nhefs9282
}
\description{
\code{nhefs9282} is data from the 1982 and 1992 waves of the \href{https://wwwn.cdc.gov/nchs/nhanes/nhefs/default.aspx}{NHANES I Epidemiologic Follow-up Study (NHEFS)}
column names that end in 1982 are from the 1982 survey and column names that end in 1992 are from the 1992 survey.

The columns are:
\itemize{
\item \code{HANESEQ} - Subject ID
\item \code{wt1982} - 1982 weight (lb)
\item \code{wt1992} - 1992 weight (lb)
\item \code{smoke1982} - Did subject smoke in 1982 (1 = Yes, 2 = No)
\item \code{smoke1992} - Did subject smoke in 1992 (1 = Yes, 2 = No)
\item \code{sex} - Sex (1 = Male, 2 = Female)
\item \code{genhealth1982} - General health (1 = Excellent, 2 = Very Good, 3 = Good, 4 = Fair, 5 = Poor)
\item \code{hyperhb1982} - Told by a doctor that you had hypertension or high blood pressure? (1 = Yes, 2 = No)
\item \code{cancer1982} - Did a doctor ever tell you that you had any cancer? (1 = Yes, 2 = No)
\item \code{cough1982} - Do you usually have a cough? Exclude clearing your throat. (1 = Yes, 2 = No)
\item \code{bedridden1982} - Are you usually confined to bed for moat of the day? (1 = Yes, 2 = No)
\item \code{asthma1982} - Did a doctor ever tell you that you had asthma? (1 = Yes, 2 = No)
\item \code{bronchemphy1982} - Did a doctor ever tell you that you had chronic bronchitis, emphysema? (1 = Yes, 2 = No)
\item \code{migraine1982} - Did a doctor ever tell you that you had migraine? (1 = Yes, 2 = No)
\item \code{heartcond1982} - Did a doctor ever tell you that you had heart condition or heart trouble? (1 = Yes, 2 = No)
\item \code{heartattack1982} - Did a doctor ever tell you that you had heart condition or heart attack? (1 = Yes, 2 = No)
\item \code{depressed1982} - During the past week I felt depressed? (1 = Yes, 2 = No)
\item \code{numcig1982} - About how many cigarettes a day do you smoke? (0 means less than one a day, 1 - 100 is number of cigarettes)
\item \code{amntalc1982} - On the average, how often do you drink alcoholic beverages, that is, beer, wine or liquor? (1-31 = Number of days, 94 = recently quit drinking, 95 = more than 3 but less than 12 times a year, no more than 3 times a year)
\item \code{drinkcoff1982} - Do you drink coffee? (1 = Yes, 2 = No)
\item \code{physicalactive1982} - In your usual day, aside from recreation, are you physically very active, moderately active, or quite inactive? (1 = very active, 2 = moderately active, 3 = quite inactive)
\item \code{marital1982} - Are you now married, widowed, divorced, separated or have you never been married? (1 = Married, 2 = Windowed, 3 = Divorced, 4 = Separated, 5 = Never Married)
\item \code{urbanrural1982} - Would you say that you have lived most of your life in a rural area, in the city, or in the city suburbs? (1 = rural, 2 = city, 3 = city suburbs, 4 = other (small town))
\item \code{famincome1982} - Which of these income groups represents your total combined family income for the past 12 months? (1 = < $3000, 2 = $3000 - $3999, 3 = $4000 - $4999, 4 = $5000 - $5999, 5 = $6000 - $6999, 6 = $7000 - $9999, 7 = $10000 - $14999, 8 = $15000 - $19999, 9 = $20000 - $24999, 10 = $25000 - $34999, 11 = $35000 - $49999, 12 = $50000 - $74999, 13 = $75000 - $100000)
, 14 = over $100000)
\item \code{age1992} - Age in years
\item \code{age1982} - Age in years
\item \code{marital1992} - Current marital status (1 = married, 2 = widowed, 3 = divorced, 4 = separated, 5 = never married)
\item \code{deathage1992} - Subject age at death
\item \code{height1982} - Height (in)
\item \code{weight1982} - Weight (lb)
\item \code{weight1992} - Weight (lb)
\item \code{wtgain} - Weight gain (lb) weight in 1992 minus weight in 1982
\item \code{qsmoke} - 1982 smokers that quit smoking between 1982 and 1992 (1 = yes, 2 = no)
\item \code{incomeclass} - Income classification based on 1982 family income ( 1 = famincome1982 >= 1 & famincome1982 <= 6, 2 = famincome1982 >= 7 & famincome1982 <= 8, 3 = famincome1982 >= 9 & famincome1982 <= 10, 4 = famincome1982 >= 11 & famincome1982 <= 12, 5 = famincome1982 >= 13)
}
}
\keyword{datasets}
