% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painstudy2.R
\docType{data}
\name{painstudy2}
\alias{painstudy2}
\title{painstudy2}
\format{
R data frame
}
\usage{
painstudy2
}
\description{
Simulated data of \href{https://pubmed.ncbi.nlm.nih.gov/28234636/}{Amundson, A.W. et al. (2017)}.
\itemize{
\item \code{pain} - pain score ( 0 - no pain to 10 - worst possible pain )
\item \code{trt} - treatment level
}
}
\keyword{datasets}
