% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtdat.R
\docType{data}
\name{rtdat}
\alias{rtdat}
\title{rtdat}
\format{
R data frame
}
\usage{
rtdat
}
\description{
Reaction time data
\itemize{
\item \code{rt} - reaction time (seconds)
\item \code{group} - treatment level
}
}
\keyword{datasets}
