% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scidesignR_example.R
\name{scidesignR_example}
\alias{scidesignR_example}
\title{Get path to scidesignR example}
\usage{
scidesignR_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
scidesignR comes bundled with a few data sets in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
scidesignR_example()
scidesignR_example("nhefshwdat.csv")
}
