% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shoedat_obs.R
\docType{data}
\name{shoedat_obs}
\alias{shoedat_obs}
\title{shoedat_obs}
\format{
R data frame
}
\usage{
shoedat_obs
}
\description{
Boys shoe data based on Section 3.2 of Box et al. (2005)
\itemize{
\item \code{boy} - Subject id
\item \code{sideA} - Side material A was assigned (L - left side, R - right side)
\item \code{sideB} - Side material B was assigned (L - left side, R - right side)
\item \code{wearA} - Amount of wear from the shoe assigned material A
\item \code{wearB} - Amount of wear from the shoe assigned material B
}
}
\keyword{datasets}
