% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silkdat.R
\docType{data}
\name{silkdat}
\alias{silkdat}
\title{silkdat}
\format{
R data frame
}
\usage{
silkdat
}
\description{
Data from \href{https://pubs.acs.org/doi/10.1021/acsbiomaterials.0c01657}{Bucciarelli, A. et a;. (2021)}
\itemize{
\item \code{Std} - Standard run number
\item \code{Run} - Randomized run number
\item \verb{A:Number of baths} - Level of number of baths
\item \code{B:Time} - Level of time
\item \code{C:Temperature} - Level of temperature
\item \code{D:Concentration} - Level of concentration
\item \code{mass_change} - Mass change (initial mass - final mass)
\item \code{mass_change_pct} - Mass percent change (initial mass - final mass)/(initial mass)
\item \code{removed_sericin_single} - Amount of removed sericin from a single bath
\item \code{removed_sericin_double} - Amount of removed sericin from a double bath
\item \code{removed_sericin_pct} - Percentage of removed sericin
}
}
\keyword{datasets}
