% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtlossdat.R
\docType{data}
\name{wtlossdat}
\alias{wtlossdat}
\title{wtlossdat}
\format{
R data frame.
}
\usage{
wtlossdat
}
\description{
\code{wtlossdat} contains data from section 5.2 of Box et al. (2005)
\itemize{
\item \code{A} - Level of food diary
\item \code{B} - Level of increasing activity
\item \code{C} - Level of home visit
\item \code{y} - Weight loss (kg)
}
}
\keyword{datasets}
